libs/deepagents-cli/pyproject.toml
[project]
name = "deepagents-cli"
version = "0.0.7"
description = "Deepagents CLI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
  "deepagents==0.2.4",
  "requests",
  "rich>=13.0.0",
  "prompt-toolkit>=3.0.52",
  "questionary>=2.0.0",  # Keep for legacy compatibility
  "cement>=3.0.0",  # New: Cement framework for CLI structure
  "langchain-anthropic>=0.1.0",
  "langchain-community>=0.1.0",
  "langchain-openai>=0.1.0",
  "tavily-python",
  "python-dotenv",
  "filelock>=3.15.0",
  "langgraph-checkpoint-sqlite>=3.0.0",
  "langgraph-checkpoint-postgres>=3.0.0",
  "psycopg[binary,pool]>=3.0.0",
  "tenacity>=8.0.0",
  "aiosqlite>=0.17.0",  # For AsyncSqliteSaver (CLI with USE_ASYNC_CHECKPOINTER=1)
]

[project.scripts]
deepagents = "deepagents_cli:cli_main"
deepagents-cli = "deepagents_cli:cli_main"

[tool.uv.sources]
deepagents = { workspace = true }

[dependency-groups]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-socket>=0.7.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.9.7",
]


dev = [
    "pytest",
    "pytest-cov",
    "build",
    "twine",
    "langchain-openai",
    "pytest-timeout>=2.4.0",
    "pytest-socket>=0.7.0",
    "pytest-asyncio>=1.2.0",
]

lint = [
    "ruff",
    "mypy",
    "types-requests"
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
deepagents_cli = ["default_agent_prompt.md"]

[tool.setuptools.packages.find]
where = ["."]
include = ["deepagents_cli*"]
exclude = ["issues*", "tests*"]

[tool.ruff]
line-length = 100
exclude = []

[tool.ruff.format]
docstring-code-format = true  # Formats code blocks in docstrings

[tool.ruff.lint]
select = [
    "ALL"  # Enable all rules by default
]
ignore = [
    "COM812",  # Messes with the formatter
    "ISC001",  # Messes with the formatter
    "PERF203", # Rarely useful
    "SLF001",  # Private member access
    "PLC0415", # Imports should be at the top. Not always desirable
    "PLR0913", # Too many arguments in function definition
    "PLC0414", # Inconsistent with how type checkers expect to be notified of intentional re-exports
    "C901",    # Too complex
    "PLR0915", # Too many statements in function
    "E501",    # Line too long (enforced by formatter for code, but docstrings can be longer)
    "ANN001",  # Missing type annotation for function argument
    "ANN201",  # Missing return type annotation
    "PTH109",  # os.getcwd() instead of Path.cwd()
    "BLE001",  # Broad except clause (legitimate in many cases)
    "S110",    # try-except-pass without logging
    "SIM105",  # Use contextlib.suppress (less readable than try/except)
    "ARG001",  # Unused function argument (common in callbacks/interfaces)
    "ARG005",  # Unused lambda argument (common in mocking)
]
unfixable = ["B028"]  # Rules that shouldn't be auto-fixed

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"  # Google-style docstrings
ignore-var-parameters = true

[tool.ruff.lint.per-file-ignores]
"deepagents_cli/cli.py" = [
    "T201",    # Allow print statements in CLI
]
"tests/*" = [
    "D1",      # Skip documentation rules in tests
    "S101",    # Allow asserts in tests
    "S311",    # Allow pseudo-random generators in tests
    "ANN201",  # Missing return type annotation
    "INP001",  # Implicit namespace package
]

[tool.mypy]
strict = false
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
disable_error_code = [
    "no-untyped-def",
    "no-untyped-call",
    "var-annotated",
    "arg-type",
    "override",
    "typeddict-item",
    "return-value",
    "name-defined",
    "call-arg",
    "assignment",
    "misc",
    "unused-ignore",
    "redundant-cast",
    "no-redef",
]
